try to create rundir and lockdir before using them
Runtime dir changed from /run to /run/pppd in commit 66a8c74c3f73 ("Let
./configure control the paths for pppd") and is likely to not exist on
some distros, in which case the pppdb will not be created.

Lockdir has a similar problem depending on the configuration, which is
even more important as failure to create a lock file leads to pppd not
starting.

Return code of mkdir does not need to be checked as the following open
will fail anyway if mkdir failed.

See: #419 (lock directory moved in ppp-2.5.0)
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>